+++ core/src/lib/services/api.mjs
@@ -23,13 +23,13 @@ export const service = {
      * We just reuse the default hook here, checking for changes in
      * name/version of the container.
      */
-    recreateContainer: (...params) => defaultRecreateContainerHook('api', ...params),
+    recreateContainer: (hookProps) => defaultRecreateContainerHook('api', hookProps),
     /**
      * Lifecycle hook to determine whether to restart the container
      * We just reuse the default hook here, checking whether the container
      * was recreated or is not running.
      */
-    restartContainer: (...params) => defaultRestartContainerHook('api', ...params),
+    restartContainer: (hookProps) => defaultRestartContainerHook('api', hookProps),
     /**
      * Lifecycle hook for anything to be done prior to creating the container
      *