+++ core/src/lib/services/broker.mjs
@@ -28,13 +28,13 @@ export const service = {
      * We just reuse the default hook here, checking for changes in
      * name/version of the container.
      */
-    recreateContainer: (...params) => defaultRecreateContainerHook('broker', ...params),
+    recreateContainer: (hookProps) => defaultRecreateContainerHook('broker', hookProps),
     /**
      * Lifecycle hook to determine whether to restart the container
      * We just reuse the default hook here, checking whether the container
      * was recreated or is not running.
      */
-    restartContainer: (...params) => defaultRestartContainerHook('broker', ...params),
+    restartContainer: (hookProps) => defaultRestartContainerHook('broker', hookProps),
     /**
      * Lifecycle hook for anything to be done prior to creating the container
      *