+++ core/src/lib/services/connector.mjs
@@ -34,13 +34,13 @@ export const service = {
      * We just reuse the default hook here, checking for changes in
      * name/version of the container.
      */
-    recreateContainer: (...params) => defaultRecreateContainerHook('connector', ...params),
+    recreateContainer: (hookProps) => defaultRecreateContainerHook('connector', hookProps),
     /**
      * Lifecycle hook to determine whether to restart the container
      * We just reuse the default hook here, checking whether the container
      * was recreated or is not running.
      */
-    restartContainer: (...params) => defaultRestartContainerHook('connector', ...params),
+    restartContainer: (hookProps) => defaultRestartContainerHook('connector', hookProps),
     /**
      * Lifecycle hook for anything to be done prior to creating the container
      *