+++ core/src/lib/services/console.mjs
@@ -24,13 +24,13 @@ export const service = {
      * We just reuse the default hook here, checking for changes in
      * name/version of the container.
      */
-    recreateContainer: (...params) => defaultRecreateContainerHook('console', ...params),
+    recreateContainer: (hookProps) => defaultRecreateContainerHook('console', hookProps),
     /**
      * Lifecycle hook to determine whether to restart the container
      * We just reuse the default hook here, checking whether the container
      * was recreated or is not running.
      */
-    restartContainer: (...params) => defaultRestartContainerHook('console', ...params),
+    restartContainer: (hookProps) => defaultRestartContainerHook('console', hookProps),
     /**
      * Lifecycle hook for anything to be done prior to creating the container
      *